fix: MiniPlayerDetailControlView 코드 역할 분리 및 테스트케이스 검증 및 수정
