--- Modules/MiniPlayer/Sources/Empty.swift
@@ -1,1 +1,0 @@
-// Empty.swift
