+++ Modules/MiniPlayer/Sources/MiniPlayerData.swift
@@ -4,6 +4,8 @@ import Combine
 
 @MainActor
 public final class MiniPlayerData: ObservableObject {
+    // MARK: - Properties
+    // MARK: Public
     @Published public private(set) var isMediaItemsFetched = false
     @Published public private(set) var miniPlayerHeight: CGFloat = 80
     @Published public private(set) var playbackStatus: PlaybackStatus = .paused
@@ -11,14 +13,16 @@ public final class MiniPlayerData: ObservableObject {
     @Published public private(set) var nowPlayTime: TimeInterval = 0
     @Published public private(set) var repeatMode: RepeatMode = .none
     
-    private var timer: Timer?
-    private var cancellables = Set<AnyCancellable>()
-    private var albumList = [MediaItemCollection]()
-    
     var title: String { nowPlayingItem?.title ?? "재생 중인 곡이 없습니다." }
     var artistName: String { nowPlayingItem?.artist ?? "Unknown Artist" }
     var artwork: Image? { nowPlayingItem?.artwork }
     
+    // MARK: Private
+    private var timer: Timer?
+    private var cancellables = Set<AnyCancellable>()
+    private var albumList = [MediaItemCollection]()
+    
+    // MARK: - Initializer
     public init() {
         Task { await notificationObservers() }
     }
@@ -34,115 +38,6 @@ public final class MiniPlayerData: ObservableObject {
         }
         
         isMediaItemsFetched = true
-    }
-    
-    // MARK: Private
-    private func updateNowPlayingItem() async {
-        self.nowPlayingItem = await MediaPlayerService.shared.nowPlayingItem()
-        self.nowPlayTime = await MediaPlayerService.shared.playbackTime()
-    }
-     
-    private func updatePlaybackState() async {
-        let playbackStatus = await MediaPlayerService.shared.playbackState()
-        self.playbackStatus = playbackStatus == .playing ? .playing : .paused
-    }
-    
-    /// 시스템 임의 재생 playbackStatus -> repeatMode:  none -> nowPlayingItem
-    private func updateQueueDidChange() async {
-        self.nowPlayingItem = await MediaPlayerService.shared.nowPlayingItem()
-    }
-    
-    /*
-     PlaybackState 노티 이후 playbackState() async -> MPMusicPlaybackState 으로 재생/정지 상태를 췍
-     NowPlayingItemDidChange 노티 이후 nowPlayingItem() async -> MediaItem? 으로 현재 곡 췍
-     
-     MPMusicPlayerControllerPlaybackStateDidChange
-     MPMusicPlayerControllerNowPlayingItemDidChange
-     MPMediaLibraryDidChange
-     MPMusicPlayerControllerQueueDidChange
-     */
-    private func notificationObservers() async {
-        await MediaPlayerService.shared.beginGeneratingPlaybackNotifications()
-        
-        /// 음악 재생 후 진입 시
-        /// VolumeDidChange -> VolumeDidChange
-        ///
-        /// 음악 재생 중 -> 음악 정지
-        /// PlaybackStateDidChange -> PlaybackStateDidChange -> PlaybackStateDidChange
-        ///
-        /// 음악 재생 중 -> 셔플
-        /// PlaybackStateDidChange -> NowPlayingItemDidChange -> QueueDidChange -> NowPlayingItemDidChange -> PlaybackStateDidChange
-        NotificationCenter.default
-            .publisher(for: .MPMusicPlayerControllerPlaybackStateDidChange)
-            .debounce(for: .milliseconds(100), scheduler: DispatchQueue.main)
-            .sink { [weak self] info in
-                Task {
-                    let playbackState = await MediaPlayerService.shared.playbackState()
-                    self?.playbackStatus = playbackState == .playing ? .playing : .paused
-                    
-                    await MainActor.run {
-                        if self?.playbackStatus == .playing {
-                            self?.startTimer()
-                        } else {
-                            self?.stopTimer()
-                        }
-                    }
-                    print("✅ playbackStatus: ", playbackState)
-                }
-            }
-            .store(in: &cancellables)
-        
-        NotificationCenter.default
-            .publisher(for: .MPMusicPlayerControllerNowPlayingItemDidChange)
-            .debounce(for: .milliseconds(100), scheduler: DispatchQueue.main)
-            .sink { [weak self] info in
-                Task {
-                    let nowPlayingItem = await MediaPlayerService.shared.nowPlayingItem()
-                    let playbackTime = await MediaPlayerService.shared.playbackTime()
-                    
-                    await MainActor.run {
-                        self?.nowPlayingItem = nowPlayingItem
-                        self?.nowPlayTime = playbackTime
-                    }
-                    print("✅ nowPlayingItem: ", nowPlayingItem)
-                    print("✅ nowPlayTime: ", playbackTime)
-                }
-            }
-            .store(in: &cancellables)
-        
-        NotificationCenter.default
-            .publisher(for: .MPMusicPlayerControllerQueueDidChange)
-            .debounce(for: .milliseconds(100), scheduler: DispatchQueue.main)
-            .sink { [weak self] info in
-                Task {
-                    let repeatMode = await MediaPlayerService.shared.repeatMode()
-                    
-                    await MainActor.run {
-                        self?.repeatMode = repeatMode
-                    }
-                    
-                    print("✅ repeatMode: ", repeatMode)
-                }
-            }
-            .store(in: &cancellables)
-    }
-    
-    private func startTimer() {
-        stopTimer()
-        timer = Timer.scheduledTimer(withTimeInterval: 1, repeats: true) { [weak self] _ in
-            Task { [weak self] in
-                guard let self = self else { return }
-                let playTime = await MediaPlayerService.shared.playbackTime()
-                await MainActor.run {
-                    self.nowPlayTime = playTime
-                }
-            }
-        }
-    }
-    
-    private func stopTimer() {
-        timer?.invalidate()
-        timer = nil
     }
     
     public func togglePlayback() async {
@@ -196,4 +91,94 @@ public final class MiniPlayerData: ObservableObject {
         return playbackStatus == .playing &&
         nowPlaying.original.albumPersistentID == album.original.representativeItem?.albumPersistentID
     }
+    
+    // MARK: Private(Data)
+    private func updateNowPlayingItem() async {
+        self.nowPlayingItem = await MediaPlayerService.shared.nowPlayingItem()
+        self.nowPlayTime = await MediaPlayerService.shared.playbackTime()
+    }
+    
+    private func updatePlaybackState() async {
+        let playbackStatus = await MediaPlayerService.shared.playbackState()
+        self.playbackStatus = playbackStatus == .playing ? .playing : .paused
+    }
+    
+    private func updateQueueDidChange() async {
+        self.nowPlayingItem = await MediaPlayerService.shared.nowPlayingItem()
+    }
+    
+    // MARK: Private(Notification)
+    private func notificationObservers() async {
+        await MediaPlayerService.shared.beginGeneratingPlaybackNotifications()
+        
+        NotificationCenter.default
+            .publisher(for: .MPMusicPlayerControllerPlaybackStateDidChange)
+            .debounce(for: .milliseconds(100), scheduler: DispatchQueue.main)
+            .sink { [weak self] info in
+                Task {
+                    let playbackState = await MediaPlayerService.shared.playbackState()
+                    self?.playbackStatus = playbackState == .playing ? .playing : .paused
+                    
+                    await MainActor.run {
+                        if self?.playbackStatus == .playing {
+                            self?.startTimer()
+                        } else {
+                            self?.stopTimer()
+                        }
+                    }
+                }
+            }
+            .store(in: &cancellables)
+        
+        NotificationCenter.default
+            .publisher(for: .MPMusicPlayerControllerNowPlayingItemDidChange)
+            .debounce(for: .milliseconds(100), scheduler: DispatchQueue.main)
+            .sink { [weak self] info in
+                Task {
+                    let nowPlayingItem = await MediaPlayerService.shared.nowPlayingItem()
+                    let playbackTime = await MediaPlayerService.shared.playbackTime()
+                    
+                    await MainActor.run {
+                        self?.nowPlayingItem = nowPlayingItem
+                        self?.nowPlayTime = playbackTime
+                    }
+                }
+            }
+            .store(in: &cancellables)
+        
+        NotificationCenter.default
+            .publisher(for: .MPMusicPlayerControllerQueueDidChange)
+            .debounce(for: .milliseconds(100), scheduler: DispatchQueue.main)
+            .sink { [weak self] info in
+                Task {
+                    let repeatMode = await MediaPlayerService.shared.repeatMode()
+                    
+                    await MainActor.run {
+                        self?.repeatMode = repeatMode
+                    }
+                }
+            }
+            .store(in: &cancellables)
+    }
+    
+    // MARK: Private(Timer)
+    private func startTimer() {
+        stopTimer()
+        timer = Timer.scheduledTimer(withTimeInterval: 1, repeats: true) { [weak self] _ in
+            Task { [weak self] in
+                guard let self = self else { return }
+                let playTime = await MediaPlayerService.shared.playbackTime()
+                await MainActor.run {
+                    self.nowPlayTime = playTime
+                }
+            }
+        }
+    }
+    
+    private func stopTimer() {
+        timer?.invalidate()
+        timer = nil
+    }
+    
+   
 }
