+++ README.md
@@ -3,6 +3,22 @@
 
 SwiftUI 기반 iOS 음악 플레이어
 
+ProjectM은 SwiftUI를 기반으로 개발된 iOS 음악 플레이어 애플리케이션 
+최신 iOS 기술 스택을 활용하여 안정적이고 직관적인 사용자 경험을 제공하며,
+모듈화된 아키텍처와 동시성 처리를 통해 유지보수성과 확장성을 높였습니다.
+ 
+ 프로젝트는 음악 재생, 앨범 관리, 미니 플레이어 컨트롤 기능을 제공하며, 
+시스템 음악 앱과의 완벽한 호환성을 목표로 개발하였습니다.
+
+**기술 스택**
+
+- SwiftUI: 선언적 UI 프레임워크를 활용하여 직관적이고 유연한 UI 구현.
+- Combine: 비동기 데이터 스트림 처리 및 노티피케이션 구독 관리.
+- Swift 6 동시성(async/await): 동시성 처리를 통해 안정적이고 효율적인 비동기 작업 지원.
+- MV Architecture: 뷰와 비즈니스 로직을 명확히 분리하여 테스트와 유지보수 용이성 확보.
+- Tuist: 프로젝트 관리 도구를 사용하여 모듈화된 구조를 효율적으로 관리.
+
+
 ## 기능 검증
 
 ### 루트 화면 기능 및 테스트 케이스
@@ -50,11 +66,31 @@ SwiftUI 기반 iOS 음악 플레이어
         - 미니 플레이어에서
             - [x]  재생/정지 버튼의 싱크가 맞는가? (해당 화면은 변경되지 않는다)
         
-    
-  // TODO: 미니플레이 상세 및 시스템 음악 앱 검증 필요  
-- 미니플레이어의 모든 기능은 시스템 음악 앱과 동기화 된다.
+### 미니플레이어 상세화면 기능 및 테스트 케이스
+`음악 앱과 동일한 기능을 제공한다`
+
+- 반복 모드를 제공한다
+    - [x]  none → all → one 으로  모드가 변경되는가
+    - [x]  시스템 앱과 상호 동기화가 되는가?
+    - [x]  각 모드 별로 앨범 내 반복모드가 작동하는가?
+- 이전 곡 / 다음 곡 재생을 제공한다
+    - [x]  이전 / 다음 곡으로 이동하는가
+    - [x]  앨범 개수를 초과하는 다음 곡인 경우 재생이 멈추는가? (시스템앱과 동일 동작)
+- 재생/정지를 제공한다
+    - [x]  시스템 앱과 앨범 상세화면과 상호 동기화가 되는가?
+- 임의 재생을 제공한다
+    - [x]  시스템 앱과 앨범 상세화면과 상호 동기화가 되는가?
+- 재생 위치 조정을 슬라이드로 제공한다
+    - [x]  슬라이드로 재생 위치를 조정할 수 있는가?
+    - [x]  재생이 끝나고 업데이트가 정상적으로 이루어 지는가?
+- 시스템 음향 조정을 슬라이드로 제공한다
+    - [x]  슬라이드로 시스템음향을 조정할 수 있는가?
+    - [x]  시스템 볼륨 조절 인디게이터가 노출되는가?
+
 
 -----------
+
+# 초기 설계 히스토리
 
 ## 기능 정의
